Evaluate Model Performance on Image Data

Given trained models and a set of images, this function evaluates the performance of the models.

Usage

evaluate_trainedmodel_performance(
  trained_models_dir,
  image_data_dir,
  output_dir = "data/",
  positive.class = "Gibbons",
  negative.class = "Noise"
)

Arguments

trained_models_dir

Path to the directory containing trained models (.pt files).

image_data_dir

Path to the directory containing image data for evaluation.

output_dir

Path to the directory where the performance scores will be saved.

positive.class

Label for the positive class.

negative.class

Label for the negative class.

Value

Invisible NULL. The performance scores are written to the specified output directory.

Examples

{
}
#> NULL