Transfer Learning from Sound Directories

This function processes sound data from a specified directory, performs transfer learning using a pre-trained deep learning model, and saves the results.

Usage

deploy_CNN_binary(
  output_folder,
  output_folder_selections,
  output_folder_wav,
  top_model_path,
  path_to_files,
  detect_pattern = NA,
  architecture,
  clip_duration = 12,
  hop_size = 6,
  downsample_rate = 16000,
  threshold = 0.5,
  save_wav = TRUE,
  positive.class = "Gibbons",
  negative.class = "Noise",
  max_freq_khz = 2
)

Arguments

output_folder

A character string specifying the path to the output folder where the results will be saved.

output_folder_selections

A character string specifying the path to the folder where selection tables will be saved.

output_folder_wav

A character string specifying the path to the folder where extracted WAV files will be saved.

top_model_path

A character string specifying the path to the pre-trained top model for classification.

path_to_files

A character string specifying the path to the directory containing sound files to process.

detect_pattern

(optional) A character string specifying a pattern to detect in the file names. Default is NA.

architecture

A character string specifying the architecture of the pre-trained model.

clip_duration

The duration of each sound clip in seconds.

hop_size

The hop size for splitting the sound clips.

downsample_rate

The downsample rate for audio in Hz. Set to 'NA' if no downsampling is required.

threshold

The threshold for audio detection.

save_wav

A logical value indicating whether to save the extracted sound clips as WAV files. Default is TRUE.

positive.class

A character string specifying the positive class label. Default is 'Gibbons'.

negative.class

A character string specifying the negative class label. Default is 'Noise'.

max_freq_khz

The maximum frequency in kHz for spectrogram visualization. Default is 2.

Details

This function processes sound data from a directory, extracts sound clips, converts them to images, performs image classification using a pre-trained deep learning model, and saves the results including selection tables and image and audio files.

Examples

{
  #' Load data
  data("TempBinWav")

  #' Create necessary directories
  dir.create(file.path(tempdir(), "BinaryDir", "Wav"), recursive = TRUE, showWarnings = TRUE)

  #' Write to temp directory
  writeWave(TempBinWav, filename = file.path(tempdir(), "BinaryDir", "Wav", "TempBinWav.wav"))

  #' Set model directory
  trained_models_dir <- system.file("extdata", "trainedresnetbinary", package = "gibbonNetR")

  #' Specify model path
  ModelPath <- list.files(trained_models_dir, full.names = TRUE)

  #' Run model
  deploy_CNN_binary(
    clip_duration = 12,
    architecture = "alexnet",
    output_folder = file.path(tempdir(), "BinaryDir", "Results", "Images"),
    output_folder_selections = file.path(tempdir(), "BinaryDir", "Results", "Selections"),
    output_folder_wav = file.path(tempdir(), "BinaryDir", "Results", "Wavs"),
    detect_pattern = NA,
    top_model_path = ModelPath,
    path_to_files = file.path(tempdir(), "BinaryDir", "Wav"),
    downsample_rate = "NA",
    threshold = 0.5,
    save_wav = FALSE,
    positive.class = "Gibbons",
    negative.class = "Noise",
    max_freq_khz = 2
  )
}
#> 1 out of 1
#> TempBinWav.wav
#> saving sound clips
#> Creating images 1 start time clips
#> Classifying images using Top Model
#> Saving output
#> Saving output
#> Saving Selection Table /tmp/RtmpmrMD5t/BinaryDir/Results/SelectionsGibbons-__TempBinWav.wavTopModelBinary.txt
#> 0.671534299850464