Evaluate Model Performance on Image Data

Given trained models and a set of images, this function evaluates the performance of the models.

Usage

evaluate_trainedmodel_performance(
  trained_models_dir,
  image_data_dir,
  output_dir = "data/",
  positive.class = "Gibbons",
  negative.class = "Noise"
)

Arguments

trained_models_dir

Path to the directory containing trained models (.pt files).

image_data_dir

Path to the directory containing image data for evaluation.

output_dir

Path to the directory where the performance scores will be saved.

positive.class

Label for the positive class.

negative.class

Label for the negative class.

Value

Invisible NULL. The performance scores are written to the specified output directory.

Examples

{
# Set directory paths for trained models and test images
trained_models_dir <- system.file("extdata", "trainedresnetbinary", package = "gibbonNetR")
image_data_dir <- system.file("extdata", "binary", "test", package = "gibbonNetR")

# Evaluate the performance of the trained models using the test images
evaluate_trainedmodel_performance(
  trained_models_dir = trained_models_dir,
  image_data_dir = image_data_dir,
  output_dir = file.path(tempdir(), 'data/'),
  positive.class = 'Gibbons',  # Label for positive class
  negative.class = 'Noise'    # Label for negative class
)

# Find the location of saved evaluation files
CSVName <- list.files(file.path(tempdir(), 'data'), recursive = TRUE, full.names = TRUE)

# Check the output of the first file
head(read.csv(CSVName[1]))

}
#> Warning: Some torch operators might not yet be implemented for the MPS device. A
#> temporary fix is to set the `PYTORCH_ENABLE_MPS_FALLBACK=1` to use the CPU as a
#> fall back for those operators:
#> ℹ Add `PYTORCH_ENABLE_MPS_FALLBACK=1` to your `.Renviron` file, for example use
#>   `usethis::edit_r_environ()`.
#> ✖ Using `Sys.setenv()` doesn't work because the env var must be set before R
#>   starts.
#>   Sensitivity Specificity Pos.Pred.Value Neg.Pred.Value Precision Recall
#> 1           1   0.0000000      0.4166667             NA 0.4166667      1
#> 2           1   0.5714286      0.6250000      1.0000000 0.6250000      1
#> 3           0   1.0000000             NA      0.5833333        NA      0
#> 4           0   1.0000000             NA      0.5833333        NA      0
#> 5           0   1.0000000             NA      0.5833333        NA      0
#> 6           0   1.0000000             NA      0.5833333        NA      0
#>          F1 Prevalence Detection.Rate Detection.Prevalence Balanced.Accuracy
#> 1 0.5882353  0.4166667      0.4166667            1.0000000         0.5000000
#> 2 0.7692308  0.4166667      0.4166667            0.6666667         0.7857143
#> 3        NA  0.4166667      0.0000000            0.0000000         0.5000000
#> 4        NA  0.4166667      0.0000000            0.0000000         0.5000000
#> 5        NA  0.4166667      0.0000000            0.0000000         0.5000000
#> 6        NA  0.4166667      0.0000000            0.0000000         0.5000000
#>   Training.Data N.epochs CNN.Architecture       AUC Threshold Frozen
#> 1          test    multi 3_resnet18_model 0.9142857       0.1   TRUE
#> 2          test    multi 3_resnet18_model 0.9142857       0.2   TRUE
#> 3          test    multi 3_resnet18_model 0.9142857       0.3   TRUE
#> 4          test    multi 3_resnet18_model 0.9142857       0.4   TRUE
#> 5          test    multi 3_resnet18_model 0.9142857       0.5   TRUE
#> 6          test    multi 3_resnet18_model 0.9142857       0.6   TRUE
#>           Class Top1Accuracy
#> 1 female.gibbon        0.375
#> 2 female.gibbon        0.375
#> 3 female.gibbon        0.375
#> 4 female.gibbon        0.375
#> 5 female.gibbon        0.375
#> 6 female.gibbon        0.375