Evaluate Model Performance on Image Data

Given trained models and a set of images, this function evaluates the performance of the models.

Usage

evaluate_trainedmodel_performance_multi(
  trained_models_dir,
  image_data_dir,
  output_dir = "data/",
  class_names,
  noise.category = "noise",
  unfreeze = TRUE
)

Arguments

trained_models_dir

Path to the directory containing trained models (.pt files).

image_data_dir

Path to the directory containing image data for evaluation.

output_dir

Path to the directory where the performance scores will be saved.

class_names

Character vector specifying class names. User specified from training data folders.

noise.category

Category label for noise class. Default is 'noise'.

unfreeze

Logical indicating whether to unfreeze model parameters. User specified based on trained model.

Value

Invisible NULL. The performance scores are written to the specified output directory.

Examples

{
  # Set directory paths for trained models and test images
  trained_models_dir <- system.file("extdata", "trainedresnetmulti", package = "gibbonNetR")
  image_data_dir <- system.file("extdata", "multiclass", "test", package = "gibbonNetR")
  class_names <- c("female.gibbon", "hornbill.helmeted", "hornbill.rhino", "long.argus", "noise")

  # Evaluate the performance of the trained models using the test images
  evaluate_trainedmodel_performance_multi(
    trained_models_dir = trained_models_dir,
    class_names = class_names,
    image_data_dir = image_data_dir,
    output_dir = file.path(tempdir(), "data/"),
    noise.category = "noise"
  )

  # Find the location of saved evaluation files
  CSVName <- list.files(file.path(tempdir(), "data"), recursive = TRUE, full.names = TRUE)

  # Check the output of the first file
  head(read.csv(CSVName[1]))
}
#> Evaluating performance of 3_resnet18_model N epochs= multi
#> female.gibbon
#> 0.1
#> 0.2
#> 0.3
#> 0.4
#> 0.5
#> 0.6
#> 0.7
#> 0.8
#> 0.9
#> 1
#> hornbill.rhino
#> 0.1
#> 0.2
#> 0.3
#> 0.4
#> 0.5
#> 0.6
#> 0.7
#> 0.8
#> 0.9
#> 1
#>   Sensitivity Specificity Pos.Pred.Value Neg.Pred.Value Precision Recall
#> 1           1         0.0      0.4166667             NA 0.4166667      1
#> 2           1         0.5      0.5882353      1.0000000 0.5882353      1
#> 3           0         1.0             NA      0.5833333        NA      0
#> 4           0         1.0             NA      0.5833333        NA      0
#> 5           0         1.0             NA      0.5833333        NA      0
#> 6           0         1.0             NA      0.5833333        NA      0
#>          F1 Prevalence Detection.Rate Detection.Prevalence Balanced.Accuracy
#> 1 0.5882353  0.4166667      0.4166667            1.0000000              0.50
#> 2 0.7407407  0.4166667      0.4166667            0.7083333              0.75
#> 3        NA  0.4166667      0.0000000            0.0000000              0.50
#> 4        NA  0.4166667      0.0000000            0.0000000              0.50
#> 5        NA  0.4166667      0.0000000            0.0000000              0.50
#> 6        NA  0.4166667      0.0000000            0.0000000              0.50
#>   Training.Data N.epochs CNN.Architecture        AUC Threshold Frozen
#> 1          test    multi 3_resnet18_model 0.07857143       0.1   TRUE
#> 2          test    multi 3_resnet18_model 0.07857143       0.2   TRUE
#> 3          test    multi 3_resnet18_model 0.07857143       0.3   TRUE
#> 4          test    multi 3_resnet18_model 0.07857143       0.4   TRUE
#> 5          test    multi 3_resnet18_model 0.07857143       0.5   TRUE
#> 6          test    multi 3_resnet18_model 0.07857143       0.6   TRUE
#>           Class Top1Accuracy
#> 1 female.gibbon    0.4166667
#> 2 female.gibbon    0.4166667
#> 3 female.gibbon    0.4166667
#> 4 female.gibbon    0.4166667
#> 5 female.gibbon    0.4166667
#> 6 female.gibbon    0.4166667